+++ experiments/reg_toy_y.txt
@@ -1,0 +1,150 @@
+-3.528068594138352720e-01
+-5.001699154630164790e-01
+-1.089834541696277359e-02
+-5.463646267319299810e-01
+4.713229670879753863e-01
+2.687819037767696373e-01
+5.381602488591514799e-01
+9.257362343840149776e-01
+8.168226199179584590e-01
+7.647170170090304975e-01
+2.914488166067151220e-01
+2.724240605369975254e-01
+5.562242724962349616e-01
+1.061318061371452348e-01
+3.462538803141850807e-01
+4.723789466884736887e-01
+1.065518418700899250e-01
+4.190996619441534898e-01
+2.344676095805887650e-01
+2.780815723419461927e-01
+2.665325708900309065e-01
+-2.467216843601053311e-01
+2.296092326842050779e-01
+1.733593786122942793e-01
+-3.956186866876660879e-01
+-5.158507831969829649e-01
+-5.789329365299135111e-01
+-5.925224726951308640e-02
+-1.248633510507900501e-01
+5.199975691148930879e-02
+-3.429330428225890737e-01
+-5.388369457053544531e-01
+-1.367472495091671403e-01
+-5.111651106218242230e-02
+1.039807103795064580e-01
+-2.827191050503786995e-02
+-2.850458669853709881e-01
+-4.263165199614094014e-02
+-4.897141188771341191e-01
+-2.251551767354116884e-01
+-9.130600241715187693e-01
+-6.815207163803190804e-01
+-7.274597145503356455e-01
+-1.041981476660264194e+00
+-8.874343174101932963e-01
+-1.409584835307334583e+00
+-3.082426795719476198e-01
+-1.280529990877250501e+00
+-1.194532403253084407e+00
+-1.506899338581725756e+00
+-1.486690484614932029e+00
+-1.612396538630699983e+00
+-1.328076389106630506e+00
+-1.714671817189854952e+00
+-1.734253960806912520e+00
+-1.489785597007031281e+00
+-1.881383879914342971e+00
+-2.179873639407796659e+00
+-1.464320986863536511e+00
+-1.061186838067486926e+00
+-1.537916726768267850e+00
+-1.486261697886270383e+00
+-1.833537349713514564e+00
+-1.300924569002712694e+00
+-1.287425321285780910e+00
+-1.423731095929041901e+00
+-1.060357751565794926e+00
+-1.408604690015274663e+00
+-2.010762340852110253e+00
+-1.733628677182523781e+00
+-1.937798722081588210e+00
+-1.611666561970649170e+00
+-1.893477582471344789e+00
+-1.282741226083486064e+00
+-1.568383073410051010e+00
+-1.885569705325689682e+00
+-1.692352291020881472e+00
+-1.080120159643418454e+00
+-1.170349202805093025e+00
+-9.399008700517803572e-01
+-1.141933400688046163e+00
+-1.467107955391485419e+00
+-9.655673166646698835e-01
+-4.829891316988546013e-01
+-5.230887238884125434e-01
+-1.250658045133189233e+00
+-5.802521555440096623e-01
+-8.232767195776666824e-01
+-3.169779472949052490e-03
+-1.092929770162019754e+00
+-4.330111613147147320e-01
+-4.931336669519273963e-01
+-3.044912894545038773e-01
+-3.672217094352703270e-01
+-1.620558453203396754e-01
+-3.475432022013796862e-01
+-6.148324552868444925e-01
+-5.713109434861765035e-01
+-5.489775172184712115e-01
+-6.946898894670678892e-01
+-7.925950201026997766e-01
+-5.531522398738326629e-01
+-7.089348035065707876e-01
+-8.158646674503584606e-01
+-3.734049729522430083e-01
+-3.097230082530649797e-01
+-6.336592147555024646e-01
+-7.357517497139146601e-01
+-7.964233465472867479e-01
+-9.273560396856980748e-02
+-2.147049026129179694e-01
+-2.762415756840576053e-01
+-4.062466053526685439e-01
+5.622209193669011618e-02
+-3.608698221997733713e-01
+3.001515814443554975e-01
+-1.508717344752095180e-01
+-1.842098304588638269e-01
+1.959989546376082747e-02
+3.416258873823192488e-01
+2.516857073063310524e-01
+1.115539362561245307e-01
+7.970476603272691474e-01
+6.446085466816730003e-01
+5.842132999677553951e-01
+8.104442937193766516e-01
+9.859515280326159870e-01
+9.386735598280415704e-01
+7.851270170388082725e-01
+1.258533960751083258e+00
+1.128001795847129118e+00
+4.036380260485290528e-01
+8.288450883863326402e-01
+8.621193187439246675e-01
+9.691804778692207911e-01
+8.993101611869641632e-01
+1.193652603466333062e+00
+7.555733725379473187e-01
+1.238132526274348333e+00
+8.818755848073973613e-01
+9.440726074298684267e-01
+1.088724501539061862e+00
+1.329615662640044382e+00
+1.100485893000485493e+00
+1.454874426950944377e+00
+1.584783387811378841e+00
+1.191798166730595909e+00
+1.460133972625317078e+00
+1.454512817219274057e+00
+7.893896692856503527e-01
