--- src/experiments/.gitignore
@@ -1,1 +1,0 @@
-results
