--- src/experiments/__init__.py
@@ -1,1 +1,0 @@
-# import src.dataset.accessor   # noqa: F401
